+++ modules/bootstrapped/test/src/smithy4s/http/HttpUriSpec.scala
@@ -34,10 +34,19 @@ final class HttpUriSpec extends FunSuite {
       "example.com",
       None,
       IndexedSeq("foo"),
-      Map("bar" -> List("2")),
+      Map(
+        "bar" -> List("2"),
+        "baz" -> List("a==2"),
+        "qux" -> List("a&b&c")
+      ),
       Option.empty
     )
     val uri = httpUri.toURI
+
+    assertEquals(
+      uri.toString,
+      "http://example.com/foo?bar=2&baz=a%3D%3D2&qux=a%26b%26c"
+    )
     assertEquals(httpUri, HttpUri.fromURI(uri))
   }