+++ modules/core/src/smithy4s/http/HttpRequest.scala
@@ -76,13 +76,13 @@ object HttpRequest {
 
     def fromHttpEndpoint[Body, I](
         httpEndpoint: HttpEndpoint[I],
-        rawHttpLabelValues: Boolean
+        smithyPathEncoding: Boolean
     ): Writer[Body, I] = new Writer[Body, I] {
       @annotation.nowarn("cat=deprecation")
       def write(request: HttpRequest[Body], input: I): HttpRequest[Body] = {
         val path =
-          if (rawHttpLabelValues) httpEndpoint.path(input)
-          else httpEndpoint.encodedPath(input)
+          if (smithyPathEncoding) httpEndpoint.encodedPath(input)
+          else httpEndpoint.path(input)
         val staticQueries = httpEndpoint.staticQueryParams
         val oldUri = request.uri
         val newUri =