+++ modules/core/src/smithy4s/http/HttpUnaryClientCodecs.scala
@@ -46,7 +46,7 @@ object HttpUnaryClientCodecs {
       requestTransformation = F.pure(_),
       responseTransformation = F.pure(_),
       hostPrefixInjection = true,
-      rawHttpLabelValues = false
+      smithyPathEncoding = false
     )
 
   trait Builder[F[_], Request, Response] {
@@ -64,7 +64,10 @@ object HttpUnaryClientCodecs {
     def withRequestTransformation[Request1](f: Request => F[Request1]): Builder[F, Request1, Response]
     def withResponseTransformation[Response0](f: Response0 => F[Response]): Builder[F, Request, Response0]
     def withHostPrefixInjection(enabled: Boolean): Builder[F, Request, Response]
+    @deprecated("use withSmithyPathEncoding instead (opposite meaning)", "0.18.41")
     def withRawHttpLabelValues(enabled: Boolean): Builder[F, Request, Response]
+    def withSmithyPathEncoding(enabled: Boolean): Builder[F, Request, Response]
+    def hasSmithyPathEncoding: Boolean
     def build(): UnaryClientCodecs.Make[F, Request, Response]
   }
 
@@ -83,7 +86,7 @@ object HttpUnaryClientCodecs {
       requestTransformation: HttpRequest[Blob] => F[Request],
       responseTransformation: Response => F[HttpResponse[Blob]],
       hostPrefixInjection: Boolean,
-      rawHttpLabelValues: Boolean
+      smithyPathEncoding: Boolean
   )(implicit F: MonadThrowLike[F])
       extends Builder[F, Request, Response] {
     def withOperationPreprocessor(fk: PolyFunction5[OperationSchema, OperationSchema]): Builder[F, Request, Response] =
@@ -116,7 +119,12 @@ object HttpUnaryClientCodecs {
 
     def withHostPrefixInjection(enabled: Boolean): Builder[F, Request, Response] = copy(hostPrefixInjection = enabled)
     override def withRawHttpLabelValues(enabled: Boolean): Builder[F, Request, Response] =
-      copy(rawHttpLabelValues = enabled)
+      withSmithyPathEncoding(!enabled)
+
+    override def withSmithyPathEncoding(enabled: Boolean): Builder[F, Request, Response] =
+      copy(smithyPathEncoding = enabled)
+
+    def hasSmithyPathEncoding: Boolean = smithyPathEncoding
 
     def build(): UnaryClientCodecs.Make[F, Request, Response] = {
       val setBody: HttpRequest.Writer[Blob, Blob] = Writer.lift((req, blob) => req.copy(body = blob))
@@ -194,7 +202,7 @@ object HttpUnaryClientCodecs {
             HttpEndpoint.cast(endpoint).toOption match {
               case Some(httpEndpoint) => {
                 val httpInputEncoder =
-                  HttpRequest.Writer.fromHttpEndpoint[Blob, I](httpEndpoint, rawHttpLabelValues)
+                  HttpRequest.Writer.fromHttpEndpoint[Blob, I](httpEndpoint, smithyPathEncoding)
                 val requestEncoder =
                   inputEncoders.fromSchema(endpoint.input, inputEncoderCache)
                 httpInputEncoder.combine(requestEncoder)