+++ modules/core/src/smithy4s/schema/Schema.scala
@@ -152,9 +152,14 @@ sealed trait Schema[A]{
   final def isUnit: Boolean = this.shapeId == ShapeId("smithy.api", "Unit")
 
   /**
-    * Turns this schema into an error schema.
+    * Turns this schema into an error schema, using a partial function.
     */
   final def error(unlift: A => Throwable)(lift: PartialFunction[Throwable, A]) : ErrorSchema[A] = ErrorSchema(this, lift.lift, unlift)
+
+  /**
+    * Turns this schema into an error schema.
+    */
+  final def error(unlift: A => Throwable)(lift: Throwable => Option[A]) : ErrorSchema[A] = ErrorSchema(this, lift, unlift)
 
 }