+++ modules/http4s/src/smithy4s/http4s/internals/SimpleRestJsonCodecs.scala
@@ -39,13 +39,13 @@ private[http4s] class SimpleRestJsonCodecs(
     val jsonCodecs: JsonPayloadCodecCompiler,
     val fieldFilter: FieldFilter,
     val hostPrefixInjection: Boolean,
-    val rawHttpLabelValues: Boolean
+    val smithyPathEncoding: Boolean
 ) extends SimpleProtocolCodecs {
   private val hintMask =
     alloy.SimpleRestJson.protocol.hintMask
 
   def transformJsonCodecs(f: JsonPayloadCodecCompiler => JsonPayloadCodecCompiler): SimpleRestJsonCodecs =
-    new SimpleRestJsonCodecs(f(jsonCodecs), fieldFilter, hostPrefixInjection, rawHttpLabelValues)
+    new SimpleRestJsonCodecs(f(jsonCodecs), fieldFilter, hostPrefixInjection, smithyPathEncoding)
 
   @deprecated(
     message = """Use withFieldFilter instead.
@@ -70,19 +70,23 @@ private[http4s] class SimpleRestJsonCodecs(
     jsonCodecs.configureJsoniterCodecCompiler(_.withFieldFilter(fieldFilter)),
     fieldFilter,
     hostPrefixInjection,
-    rawHttpLabelValues
+    smithyPathEncoding
   )
 
+  @deprecated("Use withSmithyPathEncoding instead (it has the opposite meaning to raw http label values)", "0.18.41")
   def withRawHttpLabelValues(enabled: Boolean): SimpleRestJsonCodecs =
+    withSmithyPathEncoding(!enabled)
+
+  def withSmithyPathEncoding(enabled: Boolean): SimpleRestJsonCodecs =
     new SimpleRestJsonCodecs(
       jsonCodecs = jsonCodecs,
       fieldFilter = fieldFilter,
       hostPrefixInjection = hostPrefixInjection,
-      rawHttpLabelValues = enabled
+      smithyPathEncoding = enabled
     )
 
   def withHostPrefixInjection(newHostPrefixInjection: Boolean): SimpleRestJsonCodecs =
-    new SimpleRestJsonCodecs(jsonCodecs, fieldFilter, newHostPrefixInjection, rawHttpLabelValues)
+    new SimpleRestJsonCodecs(jsonCodecs, fieldFilter, newHostPrefixInjection, smithyPathEncoding)
 
   // val mediaType = HttpMediaType("application/json")
   private val payloadEncoders: BlobEncoder.Compiler =
@@ -134,10 +138,10 @@ private[http4s] class SimpleRestJsonCodecs(
       )
       .withBaseRequest(_ => baseRequest.pure[F])
       .withRequestMediaType("application/json")
-      .withRequestTransformation(fromSmithy4sHttpRequest[F](_).pure[F])
+      .withRequestTransformation(fromSmithy4sHttpRequest[F](_, encodePathSegments = !smithyPathEncoding).pure[F])
       .withResponseTransformation[Response[F]](toSmithy4sHttpResponse[F](_))
       .withHostPrefixInjection(hostPrefixInjection)
-      .withRawHttpLabelValues(rawHttpLabelValues)
+      .withSmithyPathEncoding(smithyPathEncoding)
       .build()
 
   }